% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_height.R
\name{add_height}
\alias{add_height}
\title{Generate best guess of height for each taxa*context}
\usage{
add_height(df, context = "cell", env_prcomp, lustr, lustr_col = "ht")
}
\arguments{
\item{df}{Dataframe with context, taxa and height columns.}

\item{context}{Character. Name of columns defining context.}

\item{env_prcomp}{List. Result from call to \code{envClean::env_pca()}}

\item{lustr}{Dataframe. Lookup from \code{lifeform} to numeric height values}

\item{lustr_col}{Character. Name of column in lustr containing height
values to use for any \code{lifeform}}
}
\value{
Dataframe with cov_col removed and replaced with best guess height in
column use_height
}
\description{
Generate best guess of height for each taxa*context
}
