% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lifeform.R
\name{add_lifeform}
\alias{add_lifeform}
\title{Generate best guess of lifeform for each taxa*context}
\usage{
add_lifeform(
  df,
  context = "cell",
  lulf = tibble::tibble(lifeform = unique(df$lifeform), sort =
    1:length(unique(df$lifeform)))
)
}
\arguments{
\item{df}{Dataframe with context, taxa and lifeform columns.}

\item{context}{Character. Name of columns defining context.}

\item{lulf}{Object mapping lifeform to some form of order to break ties.}

\item{envprcomp}{Output from env_pca()}
}
\value{
Dataframe with best guess lifeform replacing original lifeform.
}
\description{
Generate best guess of lifeform for each taxa*context
}
