% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_date.R
\name{bin_date}
\alias{bin_date}
\title{Add temporal bins to a dataframe}
\usage{
bin_date(df, date_col = "date", bin = c("year", "month"))
}
\arguments{
\item{df}{Dataframe, with \code{date_col}, to be augmented with temporal bins}

\item{date_col}{Character. Name of column in \code{df} containing dates}

\item{bin}{Character or dataframe. If character, any lubridate package
method(s) such as \code{year} for \code{lubridate::year()} or \code{month} for
\code{lubridate::month()}. If dataframe, it must contain a \code{date_col} column with
a match for every instance in \code{df} \code{data_col} (or at least every instance
needed in a temporal bin).}
}
\value{
\code{df} with extra columns from \code{bin}
}
\description{
Add temporal bins to a dataframe
}
