% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_geo_rel.R
\name{bin_geo_rel}
\alias{bin_geo_rel}
\title{Add a spatial reliability column, binned to contexts}
\usage{
bin_geo_rel(
  df,
  dist_col = "rel_metres",
  dist_min = 100,
  dist_max = 250,
  context,
  over_ride_na = NULL,
  over_ride_metres = NULL
)
}
\arguments{
\item{df}{Dataframe.}

\item{dist_col}{Character. Name of the column containing the spatial
reliability.}

\item{dist_min}{Numeric. In the same units as \code{dist_col}. The target spatial
reliability that will be filtered on later in the workflow.}

\item{dist_max}{Numeric. In the same units as \code{dist_col}. In some cases,
there is good reason to believe that \code{dist_col} is an under estimate of
spatial reliability. These cases are identified by \code{over_ride_metres} where
those instances have \code{dist_col <= dist_max }. In those cases, \code{dist_col_adj}
will contain \code{dist_min} rather than \code{dist_col}. Only needed if
\code{over_ride_metres} is used}

\item{context}{Character. Column names defining the context.}

\item{over_ride_na}{Named list. List names must be the same as column names.
Any names in \code{over_ride_na} will be matched to column names in \code{df} and any
values in that list element will be given the value \code{dist_min}. This is mainly
used to prevent filtering data sources that do not have a concept equivalent
to \code{rel_metres}.}

\item{over_ride_metres}{Named list. List names must be the same as column
names. Any names in \code{over_ride_metres} will be matched to column names in \code{df}
and any values in that list element will be given the value \code{dist_min}.}
}
\value{
tibble with additional column \code{rel_metres_adj} containing the
minimum \code{rel_metres} available within that context, potentially taking into
account any over rides. Unlike \code{reduce_geo_rel()} (which only returns a
single row per context), with \code{bin_geo_rel()} the original \code{df} is only
altered by the additional column \code{rel_metres_adj}.
}
\description{
Add a spatial reliability column, binned to contexts
}
