% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_method.R
\name{bin_method}
\alias{bin_method}
\title{Add standardised \code{method} column}
\usage{
bin_method(df, methods_lu, method_col = "method", unassigned = "observed")
}
\arguments{
\item{df}{Dataframe to attribute with the standardised method from the \code{methods_lu}.}

\item{methods_lu}{Dataframe with 'method' column for the original method name and 'method_gp' column for the standardised methods.}

\item{method_col}{Character. Name of column in \code{df} with method.}

\item{unassigned}{Character. Value for unmatched or NA methods.}
}
\value{
Dataframe with standardised \code{method} column.
}
\description{
\code{method} column in returned data frame is the match for the name in
\code{method_col} based on the supplied \code{methods_lu}.
}
