% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_quotes.R
\name{clean_quotes}
\alias{clean_quotes}
\title{Remove any ' or " from specified columns in a dataframe}
\usage{
clean_quotes(df, cols = "original_name")
}
\arguments{
\item{df}{Dataframe}

\item{cols}{Character. Column from which to remove quotes}
}
\value{
Dataframe with quotes removed from \code{cols}
}
\description{
Useful due to the reliance on original_name as an 'id' throughout env
packages. If not removed, quotes are sometimes escaped by the underlying
functions used, apparently creating duplicates among original_name.
}
\examples{
x <- data.frame(a = c("species A", "species 'A'"))
clean_quotes(x, cols = "a")
}
\keyword{internal}
