% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_text.R
\name{cleaning_text}
\alias{cleaning_text}
\title{Write a sentence describing change in taxa, records, visits and sites between
two cleaning steps}
\usage{
cleaning_text(cleaning_summary)
}
\arguments{
\item{cleaning_summary}{Dataframe. Result of call to
\code{envClean::cleaning_summary()}}
}
\value{
Dataframe with added columns that can be used in .Rmd to report on
the cleaning process.
}
\description{
Write a sentence describing change in taxa, records, visits and sites between
two cleaning steps
}
