% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{create_effort_mod}
\alias{create_effort_mod}
\title{Model the effect of principal components axes on taxa richness.}
\usage{
create_effort_mod(
  df,
  env_prcomp,
  context = "cell",
  do_iter = 1000,
  do_chains = 3,
  threshold = 0.05
)
}
\arguments{
\item{df}{Dataframe. Cleaned data specifying context.}

\item{env_prcomp}{Output from env_pca.}

\item{context}{Character. Column names that define context, usually a 'visit'
to a 'cell'.}

\item{do_iter}{Numeric specifying the number of iterations to run for each
chain in rstan analysis.}

\item{do_chains}{Numeric specifying the number of chains to run in rstan
analysis}

\item{threshold}{Numeric between 0 and 1 specifying the two-tailed threshold
above/below which richness is excessively above or below 'normal' for}
}
\value{
List of model outputs.
}
\description{
Model the effect of principal components axes on taxa richness.
}
