% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{create_lifeform}
\alias{create_lifeform}
\title{Get unique lifeform across species, perhaps including further context}
\usage{
create_lifeform(
  df,
  sppcol = "Taxa",
  lfcol = "lifeform",
  context = NULL,
  luLifeformObj = luLifeform
)
}
\arguments{
\item{df}{Dataframe with species column.}

\item{sppcol}{Character. Name of species column.}

\item{lfcol}{Character. Name of lifeform (id) column.}

\item{context}{Charcter or NULL. Set of columns that define a context within
which to generate lifeform.}

\item{luLifeformObj}{Dataframe. Name of dataframe object containing columns
used later to deal with duplicates. e.g. if a taxa (within a context) has
two lifeforms with the same number of instances, say, SA and SB, information
in luLifeformObj helps to choose the tallest option.}
}
\value{
Dataframe with columns sppcol, visit col(s), lifeform
}
\description{
There are two tasks here:
}
\details{
\itemize{
\item find the most frequent lifeform assigned to a taxa (perhaps including
other context)
\item ensure there are no duplicates across the relevant context
}
}
