% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{env_pca}
\alias{env_pca}
\title{Principal components analysis and various outputs from environmental data}
\usage{
env_pca(envdf, axes = 3, cuts = 20, intstyle = "quantile")
}
\arguments{
\item{envdf}{Dataframe containing 'cell' and environmental data.}

\item{axes}{Numeric. Number of axes to return.}

\item{cuts}{Numeric. Number of cuts along pc1. pcn gets cuts/n cuts.}

\item{intstyle}{Character. Method passed to classInt::classIntervals.}
}
\value{
List of pca outputs.
}
\description{
Principal components analysis and various outputs from environmental data
}
