% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_aoi}
\alias{filter_aoi}
\title{Filter a dataframe with e/n or lat/long to an area of interest polygon (sf)}
\usage{
filter_aoi(df, aoi, x = "long", y = "lat", crsdf = 4326, crsaoi = useEPSG)
}
\arguments{
\item{df}{Dataframe. Needs coordinate columns}

\item{aoi}{sf. Name of sf object defining the area of interest}

\item{x}{Character. Name of column with x coord}

\item{y}{Character. Name of column with y coord}

\item{crsdf}{Anything that will return a legitimate crs when passed to the
crs attribute of st_transform or st_as_sf}

\item{crsaoi}{as for crsdf}
}
\value{
Dataframe filtered to area of interest
}
\description{
Filter a dataframe with e/n or lat/long to an area of interest polygon (sf)
}
