% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter any context with less instances than a threshold value}
\usage{
filter_counts(df, context, thresh = 1)
}
\arguments{
\item{df}{Dataframe with column names defining context.}

\item{context}{Character. columns defining context within which to count
instances.}

\item{thresh}{Numeric. Threshold (inclusive of thresh) below which to filter.}
}
\value{
Filtered dataframe with same names as df
}
\description{
Filter any context with less instances than a threshold value
}
