% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_geo_range.R
\name{filter_geo_range}
\alias{filter_geo_range}
\title{Filter a dataframe with e/n or lat/long to an area of interest polygon (sf)}
\usage{
filter_geo_range(df, use_aoi, x = "long", y = "lat", crs_df = 4326, crs_aoi)
}
\arguments{
\item{df}{Dataframe. Needs coordinate columns}

\item{use_aoi}{sf. Name of sf object defining the area of interest}

\item{x}{Character. Name of column with x coord}

\item{y}{Character. Name of column with y coord}

\item{crs_df}{Anything that will return a legitimate crs when passed to the
crs attribute of st_transform or st_as_sf}

\item{crs_aoi}{as for crsdf}
}
\value{
Dataframe filtered to area of interest
}
\description{
Filter a dataframe with e/n or lat/long to an area of interest polygon (sf)
}
