% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_spat_rel}
\alias{filter_spat_rel}
\title{Filter data frame to specified spatial reliability}
\usage{
filter_spat_rel(
  df,
  distCol = "maxDist",
  dist = relDist,
  visit = c("lat", "long", "dataName", "date"),
  dfRel = luRel
)
}
\arguments{
\item{df}{Dataframe.}

\item{distCol}{Character. Name of the column containing the spatial reliability.}

\item{dist}{Numeric. In units of the maxDist column.}

\item{visit}{Character. Column names defining a 'visit'.}

\item{dfRel}{Dataframe. Lookup from reliability id to reliability distance.}
}
\value{
Dataframe with records of greater than dist filtered. Filtering is done
at 'visit' level.
}
\description{
Filter data frame to specified spatial reliability
}
