% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_text_col}
\alias{filter_text_col}
\title{Filter a column annual taxa}
\usage{
filter_text_col(df, filtCol = "lifespan", filtText = "A", dfJoin = NULL)
}
\arguments{
\item{df}{Dataframe with column to filter.}

\item{filtCol}{Character. Name of column to filter.}

\item{filtText}{Character. Text(s) to filter from df.}

\item{dfJoin}{Optional dataframe. Joined to df before filter. No names from
dfJoin are returned.}
}
\value{
Filtered dataframe with same names as df
}
\description{
Filter a column annual taxa
}
