% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_visit}
\alias{filter_visit}
\title{Clean/Tidy to one row per Taxa*Visit}
\usage{
filter_visit(
  df,
  sppCol = "SPECIES",
  visit = visitCols,
  doLifeform = TRUE,
  doCover = TRUE,
  luLifeformObj = if (doLifeform) luLifeform,
  luCoverObj = if (doCover) luCover
)
}
\arguments{
\item{df}{Dataframe to reduce.}

\item{sppCol}{Character. Name of column with species.}

\item{visit}{Character. Name of columns defining a visit.}

\item{doLifeform}{Logical. If true, tidy lifeform per Taxa*Visit}

\item{doCover}{Logical. If true, tidy cover values per Taxa*Visit}

\item{luLifeformObj}{Object. Lookup to help tidy lifeform}

\item{luCoverObj}{Object. Lookup to help tidy cover}

\item{taxa}{Character. Lookup from sppCol to 'Taxa' to use. Can be generated
by taxa_taxonomy()}
}
\value{
Dataframe with columns Taxa, visit column(s) and, if chosen, lifeform
and useCover.
}
\description{
Clean/Tidy to one row per Taxa*Visit
}
