% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find local outliers}
\usage{
find_outliers(df, context, lof_minPts = 5, iqrMult = 2)
}
\arguments{
\item{df}{Dataframe with \code{context} and all other columns defining the space
in which to look for outliers.}

\item{context}{Character. Name of columns defining context.}

\item{lof_minPts}{\code{\link[dbscan:lof]{dbscan::lof()}} \code{minPts} argument.}

\item{iqrMult}{Used in \code{quantile(x, probs = 0.75) + iqrMult * IQR(x)}.
\code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}} default value is \code{1.5}.}
}
\description{
You're probably better off using \code{flag_local_outliers()} instead
}
