% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxa.R
\name{find_taxa}
\alias{find_taxa}
\title{Find how taxa changed through the cleaning/filtering/tidying process}
\usage{
find_taxa(
  taxa_to_find,
  taxa_cols = c("original_name", "taxa"),
  lookup_taxa = lutaxa
)
}
\arguments{
\item{taxa_to_find}{Character. Taxa name to find.}

\item{taxa_cols}{Character. Name of column(s) across data frames containing
taxa information.}

\item{lookup_taxa}{Dataframe with columns names matching \code{taxa_cols}.}
}
\value{

}
\description{
Find how taxa changed through the cleaning/filtering/tidying process
}
