% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_local_outliers.R
\name{flag_local_outliers}
\alias{flag_local_outliers}
\title{Find local outliers}
\usage{
flag_local_outliers(df, context, iqrMult = 2, ...)
}
\arguments{
\item{df}{Dataframe with \code{context} and all other columns defining the space
in which to look for outliers (usually environmental variables such as
climate or satellite variables)}

\item{context}{Character. Name of columns defining context.}

\item{iqrMult}{Used in \code{quantile(x, probs = 0.75) + iqrMult * IQR(x)} to set
the threshold for an outlier. e.g. \code{ggplot2::geom_boxplot()} default value is
\code{1.5}.}

\item{...}{Passed to \code{dbscan::lof()}}
}
\description{
Thin wrapper around \code{dbscan::lof()}
}
