% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flor_all}
\alias{flor_all}
\title{Example of data combined from several data sources.}
\format{
A data frame with 13102 rows and 13
variables:
\describe{
\item{lat}{Latitude in decimal degrees.}
\item{long}{Longitude in decimal degrees.}
\item{data_name}{Name of original data source.}
\item{site}{Site name from original data source.}
\item{date}{Date of observation of \code{original_name}.}
\item{original_name}{Taxonomy retrieved from \code{data_name}.}
\item{cover}{Cover estimate (numeric) retrieved from \code{data_name}.}
\item{cover_code}{Cover estimate (ordinal) retrieved from \code{data_name}.}
\item{quad_x}{Length of one quadrat edge.}
\item{quad_y}{Length of other quadrat edge.}
\item{rel_dist}{Reliability of \code{lat} and \code{long} in metres.}
\item{month}{Month component of \code{date}.}
\item{year}{Year component of \code{date}.}
...
}
}
\usage{
flor_all
}
\description{
Example resulting from munging several different data sources (\code{data_name})
into one tidy data set. This is the sort of starting point where \code{envClean}
tools can help.
}
\keyword{datasets}
