% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{gbif_tax}
\alias{gbif_tax}
\title{Get taxonomy from \href{https://www.gbif.org/dataset/d7dddbf4-2cf0-4f39-9b2a-bb099caae36c}{GBIF Backbone Taxonomy}.}
\usage{
gbif_tax(
  df,
  sppcol = 1,
  outFile = tempfile(),
  kingType = "Plantae",
  getCommon = FALSE,
  targetRank = "Species"
)
}
\arguments{
\item{df}{Dataframe with column of species names to resolve}

\item{sppcol}{Character. Name of column with species names}

\item{outFile}{Character. Path to save results to. Saving is iterative as
retrieving names can take some time.}

\item{kingType}{Character. Kingdom type (i.e. Plantae, Animalia etc.)}

\item{getCommon}{Logical. If TRUE, also get common names. Takes much longer.}

\item{targetRank}{Character. Default is 'Species'. At what level of the
taxonomic hierarchy are results desired.}
}
\value{
Dataframe of unique species names, other columns defining
taxonomic hierarchy and, optionally, common names.
}
\description{
Retrieve accepted name taxonomy and retrieve taxonomic hierarchy for a df
with a column of taxonomic names
}
\examples{
gbif_tax(data.frame(Species = "Eucalyptus viminalis"))


}
