% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{get_gbif_common}
\alias{get_gbif_common}
\title{Find common name from GBIF}
\usage{
get_gbif_common(key)
}
\arguments{
\item{key}{Numeric. gbif 'useagekey'}
}
\value{
Character. Best attempt at finding a common name for the supplied
useagekey
}
\description{
Find common name from GBIF
}
\examples{
get_gbif_common(1) # 1 is "Plantae"
}
