% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomy.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Get GBIF backbone taxonomy}
\usage{
get_taxonomy(
  df,
  taxa_col = "original_name",
  taxonomy_file = tempfile(),
  force_new = list(original_name = NULL, timediff = as.difftime(26, units = "weeks")),
  remove_taxa = c("BOLD:", "dead", "unverified", "annual herb", "annual grass", "\\\\?"),
  remove_strings = c("\\\\sx\\\\s.*", "\\\\sX\\\\s.*", "\\\\s\\\\-\\\\-\\\\s.*",
    "\\\\s\\\\(.*\\\\)", "\\\\ssp\\\\.$", "\\\\sssp\\\\.$"),
  remove_dead = FALSE,
  ...
)
}
\arguments{
\item{df}{Dataframe with taxa column.}

\item{taxa_col}{Character. Name of column with taxa names}

\item{taxonomy_file}{Character. Path to save results to.}

\item{force_new}{List with elements \code{taxa_col} and \code{difftime}. If
\code{taxonomy_file} already exists any \code{taxa_col} matches between \code{force_new} and
\code{taxonomy_file} will be requeried. Likewise any \code{original_name} that has not
been searched since \code{difftime} will be requeried. Note the name \code{taxa_col}
should be as provided as per the \code{taxa_col} argument. Set either to \code{NULL}
to ignore.}

\item{remove_taxa}{Character. Regular expressions to be matched. Any matches
will be filtered before searching. Removes any rows that match.}

\item{remove_strings}{Character. Regular expressions to be matched. Any
matches will be removed from the string before searching. Removes any
text that matches, but the row remains.}

\item{...}{Arguments passed to \code{rgbif::name_backbone_checklist()}.}
}
\value{
Dataframe. Results from \code{envClean::get_gbif_tax()}. Tweaked by column
\code{rank} being lowercase and ordered factor as per \code{envClean::lurank}. Writes
\code{taxonomy_file} and \code{gsub("\\\\.", "_accepted.", taxonomy_file)}
}
\description{
Only queries GBIF for taxa not already in \code{taxonomy_file}.
}
\details{
Common (vernacularName) no longer supported here. Use \code{get_gbif_common()} on
a downstream result. It may be helpful to keep a usageKey through the
cleaning process for use in getting common names. Part of the reason for
removing that functionality here was the ambiguity of which key to use,
particularly around species vs subspecies.
}
