% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_ctime.R
\name{last_ctime}
\alias{last_ctime}
\title{Return the object with the latest \code{ctime} attribute}
\usage{
last_ctime(prefix = "bio_")
}
\arguments{
\item{prefix}{}
}
\description{
Return the object with the latest \code{ctime} attribute
}
