% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luclean}
\alias{luclean}
\title{Dataframe of cleaning steps}
\format{
A data frame with 19 rows and 3
variables:
\describe{
\item{clean}{Character short name for cleaning step.}
\item{desc}{Phrase describing filtering step.}
\item{order}{Suggested order}
...
}
}
\usage{
luclean
}
\description{
Dataframe of cleaning steps
}
\keyword{datasets}
