% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luRank}
\alias{luRank}
\title{Dataframe of taxonomic ranks}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{Rank}{Taxonomic rank}
\item{sort}{Numeric. Can be used directly to sort taxonomic ranks but is
also used to create the ordered factor Rank.}
...
}
}
\usage{
luRank
}
\description{
A dataset listing taxonomic rank as an ordered factor.
}
\keyword{datasets}
