% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lurelBDBSA}
\alias{lurelBDBSA}
\title{Lookup for RELIABNR field from BDBSA}
\format{
A data frame with 26 rows and 2 variables:
\describe{
\item{rel}{\code{RELIABNR} field from BDBSA}
\item{max_dist}{Maximum spatial reliability distance in metres.}
...
}
}
\usage{
lurelBDBSA
}
\description{
\code{RELIABNR} field in the \href{https://www.environment.sa.gov.au/topics/Science/Information_data/Biological_databases_of_South_Australia}{Biological Databases of South Australia}
describes the reliability of the spatial coordinates associated with any
record. In \code{lurelBDBSA} the codes from \code{RELIABNR} are mapped directly to a maximum
distance in metres allowing more easy filtering of desired spatial accuracy.
}
\keyword{datasets}
