% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aoi.R
\name{make_aoi}
\alias{make_aoi}
\title{Create an 'area of interest'}
\usage{
make_aoi(
  polygons,
  filterpolys = FALSE,
  filterpolyscol = NULL,
  buffer,
  domask = TRUE,
  usecrs = 3577
)
}
\arguments{
\item{polygons}{sf.}

\item{filterpolys}{character. Used to filter \code{filterpolyscol}.}

\item{filterpolyscol}{character. Which column to filter on.}

\item{buffer}{numeric. Create a buffer around the area of interest of
\code{buffer} metres.}

\item{domask}{logical. If FALSE, just use extent of \code{buffer}.}

\item{usecrs}{numeric. \href{https://epsg.io/}{EPSG} code giving coordinate
system to use in output sf.}
}
\value{
sf
}
\description{
Create an 'area of interest'
}
