% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_attribute.R
\name{make_attribute}
\alias{make_attribute}
\title{Title}
\usage{
make_attribute(
  df,
  taxa_col = "original_name",
  att_col,
  taxonomy,
  max_guess = "family",
  agg_method = median,
  agg_round = 2,
  unknown_action = c("remove", "unknown"),
  context = NULL,
  remove_strings = c("n/a", "''", "NA"),
  ...
)
}
\arguments{
\item{df}{Dataframe with \code{taxa_col} and \code{att_col}}

\item{taxa_col}{Character name of column in \code{df} that was passed to \code{get_taxonomy} as \code{taxa_col}}

\item{att_col}{Character name of column in \code{df} that contains the attribute to summarise}

\item{taxonomy}{List resulting from call to \code{make_taxonomy()}}

\item{max_guess}{Character. If attribute is not available for \code{taxa}, try guessing from values up to \code{max_guess} level of taxonomic hierarchy. See \code{lurank}. Note it does not make sense to provide a rank here that is lower than the \code{target_rank} provided to \code{make_taxonomy} when \code{taxonomy} was made.}

\item{agg_method}{Function name to use for summarising numeric \code{att_col}. Ignored if \code{att_col} is not numeric}

\item{agg_round}{Passed to \code{base::round()} round argument. Used if summarising numeric \code{att_col}.}

\item{unknown_action}{When no other value can be found/guessed: either default 'remove' (no record returned); or provide a value to be used. Only used for character \code{att_col}}

\item{context}{Any other columns in \code{df} to maintain throughout summarising.}

\item{remove_strings}{Character. Any values in \code{att_col} to exclude}

\item{...}{Passed to \code{agg_method}}
}
\value{
Dataframe with one row for each taxa and context with best guess at a single attribute based on the values in \code{att_col}
}
\description{
Title
}
