% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_common.R
\name{make_common}
\alias{make_common}
\title{Make common name lookup}
\usage{
make_common(
  df,
  taxa_col = "original_name",
  common_col = "common",
  taxonomy = taxonomy,
  max_guess = "species",
  context = "kingdom",
  remove_strings = c("n/a", "''", "NA", "^\\\\s*$"),
  use_aoi = NULL,
  df_x = "long",
  df_y = "lat",
  crs_df = 4326
)
}
\arguments{
\item{df}{Dataframe with \code{taxa_col} and \code{common_col}, and optionally x, y coordinate columns if \code{use_aoi} == TRUE.}

\item{taxa_col}{Character name of column in \code{df} that was passed to \code{get_taxonomy} as \code{taxa_col}.}

\item{common_col}{Character name of column in \code{df} that contains the common names.}

\item{taxonomy}{List resulting from call to \code{make_taxonomy()}.}

\item{max_guess}{Character. If common name values are not available for \code{taxa}, try guessing from values up to \code{max_guess}
level of taxonomic hierarchy. See \code{lurank}. Note it does not make sense to provide a rank here that is lower than
the \code{target_rank} provided to \code{make_taxonomy} when \code{taxonomy} was made.}

\item{context}{Any other columns in \code{df} to maintain throughout summarising.}

\item{remove_strings}{Character. Any values in \code{common_col} to exclude.}

\item{use_aoi}{sf. Name of sf object for filtering data for common name generation using envClean::filter_geo_range.}

\item{df_x}{Character. Name of column with x coordinate.}

\item{df_y}{Character. Name of column with y coordinate.}

\item{crs_df}{Anything that will return a legitimate crs when passed to the
crs attribute of st_transform or st_as_sf.}
}
\value{
Dataframe with one row for each taxa with best guess at a common name based on the values in \code{common_col}.
}
\description{
Finds common names for taxa based on highest frequency of occurrence in data using envClean::make_attribute.
Compared to a straight run of envClean::make_attribute, this function first attempts to match a taxa to a
common name based on only common names belonging to taxa names of the same rank (i.e. species names will only
be matched to species level common names if they occur in the data). Taxa that cannot be matched to a common
name at their taxonomic rank will be given a common name from a lower level (e.g. a species without a species
level match will be given the most frequently occurring subspecies common name).
}
