% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_con_status.R
\name{make_con_status}
\alias{make_con_status}
\title{Make conservation status from existing status codes}
\usage{
make_con_status(
  df,
  taxa_col = "taxa",
  status_col = "epbc",
  statuses = c("ssp", "VU", "EN", "sp", "CR", "EX", "CD", "V", "E", "SP")
)
}
\arguments{
\item{df}{Dataframe with \code{taxa_col} and \code{status_col}.}

\item{taxa_col}{Character. Name of column with taxa.}

\item{status_col}{Character. Name of column with conservation status.}

\item{statuses}{Character. Possible status codes to concatenate.}
}
\value{
Dataframe with unique taxa and their conservation status.
}
\description{
Useful for assigning a 'threatened' status to a binomial where different
statuses apply at trinomial level. The resulting 'status' is simply the
concatenated values from lower levels of the taxonomic hierarchy.
}
