% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cover.R
\name{make_cover}
\alias{make_cover}
\title{Make a single (numeric, proportion) cover column from different sorts of
input cover}
\usage{
make_cover(
  df,
  taxa_col = "taxa",
  context = NULL,
  lucov,
  cov_type = "cover_mid"
)
}
\arguments{
\item{df}{Dataframe containing columns to consolidate.}

\item{taxa_col}{Character. Name of column containing taxa.}

\item{context}{Character. Name of columns defining the context.}

\item{lucov}{Dataframe. Lookup from \code{cover_code} to numeric cover values.}

\item{cov_type}{Character. Name of column in \code{lucov} that is used to convert
character \code{cover_code} to numeric values.}
}
\value{
Dataframe with consolidated \code{use_cover} column.
}
\description{
Assumes a numeric (percentage) cover column called 'cover' and character
column called 'cover_code' that is the modified Braun-Blanquet
\insertCite{RN4265}{envEcosystems} cover value using
\href{https://www.environment.sa.gov.au/topics/Science/Information_data/Biological_databases_of_South_Australia}{Biological Databases of South Australia}
codes (\code{COVCODE} field in BDBSA). Example of \code{lucov} at
\code{envEcosystems::lucover}.
}
\references{
\insertAllCited{}
}
