% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_effort_mod.R
\name{make_effort_mod}
\alias{make_effort_mod}
\title{Distribution of credible values for taxa richness.}
\usage{
make_effort_mod(
  df,
  context = "cell",
  cat_cols = NULL,
  threshold_lo = 0.05/2,
  threshold_hi = 0.05/2,
  use_family = rstanarm::neg_binomial_2(),
  ...
)
}
\arguments{
\item{df}{Dataframe. Cleaned data specifying context.}

\item{context}{Character. Column names that define context, usually a 'visit'
to a 'cell'.}

\item{cat_cols}{Character. Name of column(s) (0 to 2) specifying the
categorical variables to model. Usually a taxonomic level (say, class) and a
geographic level (say, IBRA Region). If NULL (default), model is \code{y ~ 1}.}

\item{threshold_lo, threshold_hi}{Numeric between 0 and 1 specifying the
threshold above/below which richness is excessively above or below 'normal'
and should be filtered.}

\item{use_family}{Passed to \code{rstanarm::stan_glm} \code{family} argument.}

\item{...}{Other arguments passed to \code{rstanarm::stan_glm} (e.g. chains, iter).}
}
\value{
List of model outputs:
\item{dat_exp}{dataframe of data used in pre-model data exploration}
\item{mod}{model object}
\item{mod_pred}{dataframe resulting from rstanarm::posterior_predict}
\item{mod_resid}{dataframe of residuals}
\item{mod_resid_plot}{plot of residuals (ggplot object)}
\item{mod_res}{dataframe of summarised \code{mod_pred} results}
\item{mod_plot}{plot of distribution of credible values, faceted by any
variables in the model}
\item{mod_cell_result}{dataframe of all contexts with column \code{keep}
indicating whether the context is outside acceptable taxa richness}
\item{mod_cell_tab}{dataframe tabulating how many contexts were above and
below the acceptable richness quantile thresholds}
}
\description{
Model the effect of between 0 and 2 categorical variables on taxa richness
within a context. Highlight, based on thresholds, contexts where taxa
richness is 'excessively' high or low.
}
