% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{make_effort_mod_pca}
\alias{make_effort_mod_pca}
\title{Model the effect of principal components axes on taxa richness.}
\usage{
make_effort_mod_pca(
  df,
  env_prcomp,
  context = "cell",
  threshold_lo = 0.05/2,
  threshold_hi = 0.05/2,
  effort_col = "qsize",
  effort_thresh = 3 * 3,
  ...
)
}
\arguments{
\item{df}{Dataframe. Cleaned data specifying context.}

\item{env_prcomp}{Output from env_pca.}

\item{context}{Character. Column names that define context, usually a 'visit'
to a 'cell'.}

\item{threshold_lo, threshold_hi}{Numeric between 0 and 1 specifying the
threshold above/below which richness is excessively above or below 'normal'
and should be filtered.}

\item{effort_col}{Character. Name of column with some measure of effort.}

\item{effort_thresh}{Numeric. \code{effort_col} will be filtered below this
threshold.}

\item{...}{Passed to \code{\link[rstanarm:stan_glm]{rstanarm::stan_glm()}}.}
}
\value{
List of model outputs.
}
\description{
Model the effect of principal components axes on taxa richness.
}
