% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{make_env_pca}
\alias{make_env_pca}
\title{Principal components analysis and various outputs from environmental data}
\usage{
make_env_pca(
  env_df,
  context = "cell",
  axes = 3,
  cuts = 20,
  int_style = "quantile",
  out_file = NULL
)
}
\arguments{
\item{env_df}{Dataframe containing 'cell' and environmental data.}

\item{context}{Character. Name of columns defining context.}

\item{axes}{Numeric. Number of axes to return.}

\item{cuts}{Numeric. Number of cuts along pc1. pcn gets cuts/n cuts.}

\item{int_style}{Character. Method passed to classInt::classIntervals.}

\item{out_file}{Character. Optional path to save output.}
}
\value{
List of pca outputs.
}
\description{
Principal components analysis and various outputs from environmental data
}
