% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gbif_taxonomy.R
\name{make_gbif_taxonomy}
\alias{make_gbif_taxonomy}
\title{Make taxonomy lookups}
\usage{
make_gbif_taxonomy(
  df,
  taxa_col = "original_name",
  taxonomy_file,
  target_rank = "species",
  limit = TRUE,
  fixes = NULL,
  overrides = NULL,
  ...
)
}
\arguments{
\item{df}{Dataframe with \code{taxa_col}.}

\item{taxa_col}{Character. Name of column in \code{df} with taxa names}

\item{taxonomy_file}{Character. Path to results from
\code{envClean::get_taxonomy()}}

\item{target_rank}{Character. Default is 'species'. At what level of the
taxonomic hierarchy are results desired. This is the most detailed taxonomy
returned. i.e. if genus is the \code{target_rank}, no taxa below genus are
returned. See \code{envClean::lurank} \code{rank} column.}

\item{limit}{Logical. If true (default), the output taxonomy will be limited
to the input names in \code{df}. Otherwise, all taxa found in \code{taxonomy_file} will
be returned.}

\item{fixes}{Data frame with columns \code{resolved_to} and \code{prefer}. Any \code{taxa}
result in \code{lutaxa} that matches a name in \code{resolved_to} will be changed to
\code{prefer}. Mainly used where legitimate names are used in areas where they do
not exist. e.g. Eastern osprey \emph{Pandion cristatus} does not occur in South
Australia but records of this species in South Australia are assumed to be
legitimate Osprey (\emph{Pandion haliaetus}) records.}

\item{overrides}{Data frame with columns \code{original} and \code{prefer}. Any
\code{original_name} result in \code{lutaxa} that matches a name in \code{original} will be
have its corresponding \code{taxa} changed to \code{prefer}. Useful where GBIF Backbone
Taxonomy provides a spurious result. e.g. The GBIF Backbone Taxonomy changes
\emph{Thinornis rubricollis} to \emph{Phalaropus lobatus} rather than the preferred
\emph{Charadrius cucullatus}.}

\item{...}{Passed to \code{envClean::get_taxonomy()}}
}
\value{
named list with elements:
\item{lutaxa}{Dataframe. For each unique name in \code{taxa_col}, the best
\code{taxa} to use (taking into account \code{target_rank})}
\item{taxonomy}{Dataframe. For each \code{taxa} in \code{lutaxa} a row of taxonomic
hierarchy and matching gbif usageKeys}
}
\description{
Make taxonomy lookups
}
