% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ind.R
\name{make_ind}
\alias{make_ind}
\title{Make indigenous status lookup}
\usage{
make_ind(
  df,
  taxa_col = "original_name",
  ind_col = "ind",
  taxonomy,
  max_guess = "species",
  context = "kingdom",
  remove_strings = c("n/a", "''", "NA", "^\\\\s*$"),
  primary_data_source = NULL,
  data_source_col = NULL,
  non_ind_terms = NULL,
  common_df = NULL,
  genus_overrides = NULL,
  species_overrides = NULL,
  use_aoi = NULL,
  df_x = "long",
  df_y = "lat",
  crs_df = 4326
)
}
\arguments{
\item{df}{Dataframe with \code{taxa_col} and \code{ind_col}, and optionally x, y coordinate columns if \code{use_aoi} == TRUE.}

\item{taxa_col}{Character name of column in \code{df} that was passed to \code{get_taxonomy} as \code{taxa_col}.}

\item{ind_col}{Character name of column in \code{df} that contains the common names.}

\item{taxonomy}{List resulting from call to \code{make_taxonomy()}.}

\item{max_guess}{Character. If indigenous values are not available for \code{taxa}, try guessing from values up to \code{max_guess}
level of taxonomic hierarchy. See \code{lurank}. Note it does not make sense to provide a rank here that is lower than
the \code{target_rank} provided to \code{make_taxonomy} when \code{taxonomy} was made.}

\item{context}{Any other columns in \code{df} to maintain throughout summarising.}

\item{remove_strings}{Character. Any values in \code{ind_col} to exclude.}

\item{primary_data_source}{Character value from \code{data_source_col} indicating the name of the primary data source.}

\item{data_source_col}{Character name of the data source column with the name of the primary data source.}

\item{non_ind_terms}{Character vector of non-indigenous terms found in common names to use for applying non-indigenous
status regardless of the indigenous values in the data.}

\item{common_df}{Data frame containing \code{taxa_col} and 'common' field with common names for each taxa to use for
applying non-indigenous status with the \code{non_ind_terms}.}

\item{genus_overrides}{Character vector of known non-indigenous genera to apply non-indigenous status to all
species within those genera regardless of the indigenous values in the data.}

\item{species_overrides}{Character vector of known non-indigenous species to apply non-indigenous status regardless
of the indigenous values in the data.}

\item{use_aoi}{sf. Name of sf object for filtering data for indigenous status generation using envClean::filter_geo_range.}

\item{df_x}{Character. Name of column with x coordinate.}

\item{df_y}{Character. Name of column with y coordinate.}

\item{crs_df}{Anything that will return a legitimate crs when passed to the
crs attribute of st_transform or st_as_sf.}
}
\value{
Dataframe with one row for each taxa with best guess at a common name based on the values in \code{ind_col}.
}
\description{
Finds indigenous status for taxa based on highest frequency of occurrence in data using envClean::make_attribute.
Compared to a straight run of envClean::make_attribute, this function first attempts to find an indigenous status
for a taxa based on values found in a primary data source (if supplied). Taxa with no indigenous values in the primary data
source are then given an indigenous status based on all other data sources. In addition, flags and overrides indicating
non-indigenous taxa can also be provided to overcome errors in the data, and the indigenous calculation can be
restricted to an area of interest (aoi).
}
