% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{make_ind_status}
\alias{make_ind_status}
\title{Make indigenous status lookup}
\usage{
make_ind_status(df, taxa_col = "taxa", ind_col = "ind")
}
\arguments{
\item{df}{Dataframe with species column and column indicating indigenous
status.}

\item{taxa_col}{Character. Name of column with species.}

\item{ind_col}{Character. Name of column with indigenous status}
}
\value{
Dataframe with unique species and their indigenous status.
}
\description{
Make indigenous status lookup
}
