% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lifeform.R
\name{make_lifeform}
\alias{make_lifeform}
\title{Get unique lifeform across taxa, perhaps including further context}
\usage{
make_lifeform(
  df,
  taxa_col = "taxa",
  lf_col = "lifeform",
  context = NULL,
  lulife
)
}
\arguments{
\item{df}{Dataframe with taxa column.}

\item{taxa_col}{Character. Name of taxa column.}

\item{lf_col}{Character. Name of lifeform (id) column.}

\item{context}{Charcter or NULL. Set of columns that define a context within
which to generate lifeform.}

\item{lulife}{Dataframe lookup for lifeform.}
}
\value{
Dataframe with columns taxa_col, visit col(s), lifeform
}
\description{
There are two tasks here:
}
\details{
\itemize{
\item find the most frequent lifeform assigned to a taxa (perhaps including
other context)
\item ensure there are no duplicates across the relevant context
}
}
