% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_methods_lu.R
\name{make_methods_lu}
\alias{make_methods_lu}
\title{Make standardised collection methods lookup}
\usage{
make_methods_lu(
  df,
  methods_key,
  method_col = "method",
  taxonomy,
  taxa_col = "original_name",
  unassigned = "observed"
)
}
\arguments{
\item{df}{Dataframe with methods variants.}

\item{methods_key}{Dataframe with 'method_gp' column for the standardised methods, 'terms' column for the search terms
used to define each standard method_gp, and 'rank' column with integers specifying which method_gp to use over others
if there is overlap in search terms (where higher numbers indicate those to use over lower numbers).}

\item{method_col}{Character. Name of column in \code{df} with method.}

\item{taxonomy}{List resulting from call to \code{make_taxonomy()}.}

\item{taxa_col}{Character. Name of column in \code{df} with taxa. Used for matching to \code{taxonomy} and generating 'tax_gp'
column in output, indicating which taxonomic groups each method has been recorded for.}
}
\value{
Dataframe with original 'method' column and standardised 'method_gp' column based on the supplied \code{methods_key}.
Also, 'tax_gp' column indicating the taxonomic group for which the methods were recorded, and
'method_n' column indicating the number of occurrences for the original 'method' in the 'tax_gp'.
}
\description{
Make standardised collection methods lookup
}
