% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{make_taxa_taxonomy}
\alias{make_taxa_taxonomy}
\title{Make taxonomy lookups}
\usage{
make_taxa_taxonomy(
  df,
  taxa_col = "original_name",
  lifespan_col = NULL,
  ind_col = NULL,
  ...
)
}
\arguments{
\item{df}{Dataframe with taxa column.}

\item{taxa_col}{Name of column with taxa.}

\item{lifespan_col}{Character. Optional name of column containing lifespan
information.}

\item{ind_col}{Character. Optional name of columns containing indigenous
status of taxa in taxa_col.}

\item{...}{Passed to \code{envClean::get_gbif_tax}}
}
\value{
named list with elements:
\item{lutaxa}{dataframe. A lookup from unique values in \code{taxa_col} to
matched taxonomy from GBIF backbone}
\item{taxa_taxonomy}{dataframe with unique taxa and associated taxonomic
information}
}
\description{
Make taxonomy lookups
}
\examples{

}
