% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxonomy.R
\name{make_taxonomy}
\alias{make_taxonomy}
\title{Make taxonomy lookups}
\usage{
make_taxonomy(
  df,
  taxa_col = "original_name",
  lifespan_col = NULL,
  ind_col = NULL,
  out_file = tempfile(),
  king_type = "Plantae",
  target_rank = "species",
  remove_taxa = c("BOLD:.*\\\\d{4}", "dead", "unverified", "annual herb",
    "annual grass", "[[:alpha:]]{1}\\\\d{5,}", "^[[:alpha:]]{1,3}$"),
  remove_strings = c("\\\\s*\\\\(.*\\\\)", "\\\\'", "\\\\?", " spp\\\\.", "ssp\\\\.",
    " sp\\\\.", " var\\\\.", "subsp\\\\.", " [A-Z].*", "#", "\\\\s^", " x .*$| X .*$"),
  do_common = TRUE
)
}
\arguments{
\item{df}{Dataframe with taxa column.}

\item{lifespan_col}{Character. Optional name of column containing lifespan
information.}

\item{ind_col}{Character. Optional name of columns containing indigenous
status of taxa in taxa_col.}

\item{do_common}{Logical. If true, an attempt will be made to find a common
name. Common name matches are attempted at \code{target_rank} (or higher taxonomic
rank if no taxa was matched at \code{target_rank} for the name supplied in
\code{taxa_col}). \code{target_rank} is supplied in \code{...} to\code{envClean::get_gbif_tax()}.
\code{envClean::get_gbif_tax}.}

\item{...}{Passed to \code{envClean::get_gbif_tax()}}
}
\value{
named list with elements:
\item{raw}{Dataframe. Results from \code{envClean::get_gbif_tax()}}
\item{lutaxa}{dataframe. A lookup from unique values in \code{taxa_col} to
matched taxonomy from GBIF backbone}
\item{taxonomy}{dataframe with unique taxa and associated taxonomic
information}
\item{common}{If \code{do_common} a dataframe with unique taxa and associated
common name}
}
\description{
Make taxonomy lookups
}
\examples{

}
