% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unmatched_overrides.R
\name{make_unmatched_overrides}
\alias{make_unmatched_overrides}
\title{Attempt to find a taxa for names with no match in \code{galah::search_taxa()}}
\usage{
make_unmatched_overrides(
  df,
  taxa_col = "original_name",
  taxonomy,
  target_rank = "species",
  hybrids = FALSE,
  include_unmatched = TRUE,
  remove_taxa = c("bold:", "unverified", "undetermined", "unidentified", "annual herb",
    "annual grass", "incertae sedis", "\\\\?", "another\\\\s", "not naturalised in sa",
    "annual tussock grass", "*no id", "spec\\\\.", "\\\\s\\\\-\\\\-\\\\s.*",
    "\\\\ssp\\\\.", "\\\\sspec\\\\.", "\\\\ssp$", "\\\\ssp\\\\d", "\\\\ssp\\\\s",
    "\\\\sspp\\\\.", "\\\\sspp\\\\s", "\\\\sspp$", "dead", "unknown", "\\\\sgroup$",
    "\\\\sspecies$", "aquatic grass", "hybrid", "\\\\scultivar$", "\\\\scomplex$",
    "\\\\ssect\\\\.", "\\\\ss\\\\.\\\\sstr\\\\.", "\\\\(includes\\\\s"),
  tri_strings = c("\\\\sssp\\\\s", "\\\\sssp\\\\.", "\\\\svar\\\\s", "\\\\svar\\\\.",
    "\\\\ssubsp\\\\.", "\\\\ssubspecies", "\\\\sform\\\\)", "\\\\sform\\\\s",
    "\\\\sf\\\\.", "\\\\srace\\\\s", "\\\\srace\\\\)", "\\\\sp\\\\.v\\\\.")
)
}
\arguments{
\item{df}{Dataframe of biological records}

\item{taxa_col}{Character. Name of column in \code{df} containing the taxonomic
entities for which a match is desired.}

\item{taxonomy}{Result of call to \code{make_taxonomy()}}

\item{target_rank}{Character. Level within \code{envClean::lurank$rank} to target}

\item{hybrids}{Logical. Create overrides for hybrids (e.g. original names with 'x')?}

\item{include_unmatched}{Logical. Create overrides for taxa not matched via gbif using their original names?}

\item{remove_taxa}{Character. Taxa with regular expressions in \code{tolower(taxa_col)} that match \code{remove_taxa}
will not be searched or have overrides constructed.}

\item{tri_strings}{Character. Taxa names with these strings that indicate a trinomial will not be included
as a binomial override (i.e. avoids the use_species column in the overrides being populated with trinomial names).}
}
\value{
Tibble in appropriate form to pass to the overrides argument of
\code{make_taxonomy()}
}
\description{
For an unmatched 'name', try to find a match via: \code{rgbif::name_usage()};
\code{rgbif::name_backbone()}; and removal of any characters in 'name' after an
'x' or 'X' (i.e. treat hybrids as just the first taxa). Using
\code{rgbif::name_backbone()} allows fuzzy matching to fix spelling errors. Any
results are passed back to \code{galah::search_taxa()} to retrieve an
override to use for that 'name' (so long as the rgbif result is not identical
to 'name'). Any 'name' still completely unmatched is just given the override
'name' so it will not be lost from downstream processes but will not have any
associated taxonomic information.
}
\examples{

  # setup
  # library("envClean")

  temp_file <- tempfile()

  taxa_df <- tibble::tibble(taxa = c("Charadrius rubricollis"
                                     , "Thinornis cucullatus"
                                     , "Melithreptus gularis laetior"
                                     , "Melithreptus gularis gularis"
                                     , "Eucalyptus viminalis"
                                     , "Eucalyptus viminalis cygnetensis"
                                     , "Eucalyptus"
                                     , "Charadrius mongolus all subspecies"
                                     , "Bettongia lesueur Barrow and Boodie Islands subspecies"
                                     , "Lagorchestes hirsutus Central Australian subspecies"
                                     , "Perameles gunnii Victorian subspecies"
                                     , "Pterostylis sp. Rock ledges (pl. 185, Bates & Weber 1990)"
                                     , "Spyridium glabrisepalum"
                                     , "Spyridium eriocephalum var. glabrisepalum"
                                     , "Petrogale lateralis (MacDonnell Ranges race)"
                                     , "Gehyra montium (revised)"
                                     , "Korthalsella japonica f. japonica"
                                     , "Galaxias sp. nov. 'Hunter'"
                                     , "Some rubbish"
                                     , "Senna artemisioides subsp x artemisioides"
                                     , "Halosarcia sp.  (NC)"
                                     , "TERMITOIDAE sp." # 'epifamily'
                                     )
                            )

  # make taxonomy (returns list and writes taxonomy_file)
  taxonomy <- make_taxonomy(df = taxa_df
                            , taxa_col = "taxa"
                            , taxonomy_file = temp_file
                            , needed_ranks = c("kingdom", "genus", "species", "subspecies")
                            )
  taxonomy$raw
  taxonomy$kingdom
  taxonomy$genus
  taxonomy$species
  taxonomy$subspecies

  # query more taxa (results are added to taxonomy_file but only the new taxa are returned (default `limit = TRUE`)
  more_taxa <- tibble::tibble(original_name = c("Amytornis whitei"
                                                , "Amytornis striatus"
                                                , "Amytornis modestus (North, 1902)"
                                                , "Amytornis modestus modestus"
                                                , "Amytornis modestus cowarie"
                                                )
                              )

  make_taxonomy(df = more_taxa
                , taxonomy_file = temp_file
                , needed_ranks = c("species")
                )

  # no dataframe supplied - all results in taxonomy_file returned
  make_taxonomy(taxonomy_file = temp_file
                , needed_ranks = c("subspecies")
                )

  # Try automatic overrides
  auto_overrides <- make_unmatched_overrides(df = taxa_df
                                             , taxa_col = "taxa"
                                             , taxonomy = taxonomy
                                             , target_rank = "species"
                                             )

  # overrrides
  overrides <- envClean::taxonomy_overrides

  # C. rubricollis binned to Phalarope lobatus at species level!
  taxonomy <- make_taxonomy(df = overrides
                            , taxonomy_file = temp_file
                            , needed_ranks = c("species", "subspecies")
                            )

  taxonomy$species$lutaxa \%>\%
    dplyr::filter(grepl("rubricollis", original_name))

  # add in override - C. rubricollis is binned to T. cucullatus at species level
  taxonomy <- make_taxonomy(df = overrides
                            , taxonomy_file = temp_file
                            , needed_ranks = c("species", "subspecies")
                            , overrides = overrides
                            )

  taxonomy$species$lutaxa \%>\%
    dplyr::filter(grepl("rubricollis", original_name))


  # tweak_species example
  make_taxonomy(df = tibble::tibble(original_name = "Acacia sp. Small Red-leaved Wattle (J.B.Williams 95033)")
                , tweak_species = FALSE
                )$raw \%>\%
    dplyr::select(original_name, scientific_name, species)

  make_taxonomy(df = tibble::tibble(original_name = "Acacia sp. Small Red-leaved Wattle (J.B.Williams 95033)")
                , tweak_species = TRUE
                )$raw \%>\%
    dplyr::select(original_name, scientific_name, species)

  # clean up
  rm(taxonomy)
  unlist(paste0(temp_file, ".parquet"))
}
