% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_vis_sit_tax.R
\name{rec_vis_sit_tax}
\alias{rec_vis_sit_tax}
\title{How many records, visits, sites and taxa in a dataframe}
\usage{
rec_vis_sit_tax(df, site_cols, visit_cols, taxa_cols)
}
\arguments{
\item{df}{Dataframe with taxa records.}

\item{site_cols}{Name of column(s) defining sites.}

\item{visit_cols}{Name of column(s) defining visits.}

\item{taxa_cols}{Name of column(s) defining taxa.}
}
\value{
One row dataframe
}
\description{
How many records, visits, sites and taxa in a dataframe
}
