% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_geo_rel.R
\name{reduce_geo_rel}
\alias{reduce_geo_rel}
\title{Reduce data frame to a single spatial reliability within a context}
\usage{
reduce_geo_rel(
  df,
  dist_col = "rel_metres",
  dist_min = 100,
  dist_max = 250,
  context,
  extra_cols = "taxa",
  over_ride_na = NULL,
  over_ride_metres = NULL
)
}
\arguments{
\item{df}{Dataframe.}

\item{dist_col}{Character. Name of the column containing the spatial
reliability.}

\item{dist_min}{Numeric. In the same units as \code{dist_col}. The target spatial
reliability that will be filtered on later.}

\item{dist_max}{Numeric. In the same units as \code{dist_col}. In some cases,
there is good reason to believe that \code{dist_col} is an under estimate of
spatial reliability. These cases are identified by \code{over_ride_metres} where
those instances have \code{dist_col <= dist_max }. In those cases, \code{dist_col_adj}
will contain \code{dist_min} rather than \code{dist_col}. Only needed if
\code{over_ride_metres} is used}

\item{context}{Character. column names defining the context.}

\item{extra_cols}{Character. Extra columns added to context.}

\item{over_ride_na}{Named list. List names must be the same as column names.
Any names in \code{over_ride_na} will be matched to column names in \code{df} and any
values in that list element will be given the value \code{dist_min}. This is mainly
used to prevent filtering data sources that do not have a concept equivalent
to \code{rel_metres}.}

\item{over_ride_metres}{Named list. List names must be the same as column
names. Any names in \code{over_ride_metres} will be matched to column names in \code{df}
and any values in that list element will be given the value \code{dist_min}.}
}
\value{
dataframe with a single \code{rel_metres_adj} value per context.
}
\description{
Reduce data frame to a single spatial reliability within a context
}
