% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_clean.R
\name{report_clean}
\alias{report_clean}
\title{Knit child documents to report the standard envCleaned process}
\usage{
report_clean(clean_summary, heading_level = 2)
}
\arguments{
\item{clean_summary}{Tibble containing names of clean process outputs (e.g. clean_start, bio_bin,
bio_clean, etc), As output from envTargets::summarise_store_object}

\item{heading_level}{What level of subheading to set the output sections. E.g. default 2 will use
"##" to separate sections}
}
\description{
Use with knitr chunk option \code{results="asis"}.
Note: some files include named R objects to output the currently-used settings, and the knit will
error if these are missing. For example, the 'core' clean files require \code{scale} to describe
extent, and  \code{settings_clean} for the fbd_dist setting. Set these in the report environment before using \code{report_clean}.
}
