% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjack.R
\name{rjack}
\alias{rjack}
\title{Reverse jackknife}
\usage{
rjack(d, use_diva = TRUE)
}
\arguments{
\item{d}{Vector of values to search for outliers}

\item{use_diva}{Logical.}
}
\value{
Indices of (reverse jackknife) outliers
}
\description{
modified from https://github.com/cran/biogeo/blob/master/R/rjack.R
also see https://code.google.com/archive/p/ala-dataquality/wikis/DETECTED_OUTLIER_JACKKNIFE.wiki (including diva implementation)
}
\examples{
x <- c(rnorm(10), 2)
rjack(x)
}
\keyword{internal}
