% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{taxa_lifeform}
\alias{taxa_lifeform}
\title{Get unique lifeform across species and perhaps other columns}
\usage{
taxa_lifeform(
  df,
  sppCol = "Taxa",
  lfCol = "lifeform",
  visit = NULL,
  luLifeformObj = luLifeform
)
}
\arguments{
\item{df}{Dataframe with species column.}

\item{sppCol}{Character. Name of species column.}

\item{lfCol}{Character. Name of lifeform (id) column.}

\item{visit}{Charcter or NULL. Set of columns that define a visit.}

\item{luLifeformObj}{Dataframe. Name of dataframe object containing columns
used later to filter duplicates.}
}
\value{
Dataframe with columns sppCol, visit col(s), lifeform
}
\description{
There are two tasks here:
}
\details{
\itemize{
\item find the most frequent lifeform assigned to a taxa (perhaps including
other context)
\item ensure there are no duplicates across the relevant context
}
}
