% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{taxa_taxonomy}
\alias{taxa_taxonomy}
\title{Create taxonomy lookups}
\usage{
taxa_taxonomy(
  df,
  sppcol = "SPECIES",
  lifespancol = NULL,
  indcol = NULL,
  poorfilt = c("dead", "unverified"),
  savefile = tempfile(),
  king = "Plantae"
)
}
\arguments{
\item{df}{Dataframe with species column.}

\item{sppcol}{Name of column with species.}

\item{lifespancol}{Character. Optional name of column containing lifespan
information.}

\item{indcol}{Character. Optional name of columns containing indigenous
status of taxa in sppcol.}

\item{poorfilt}{Character. Any taxa names to grep out of the species column.
(e.g. c("annual form", "unverified")).}

\item{savefile}{Character. Path to file containing desired taxonomy to use.
This is usually the output from gfbif_tax(). If this does not exist it will
be created (as tempfile) by gbif_tax().}

\item{king}{Character. Kingdom to search preferentially in GBIF Taxonomy
Backbone}
}
\value{
Dataframe with applied taxonomy from GBIF Taxonony Backbone. Also,
two dataframes are returned to the global environment. One, named
taxaTaxonomy, with unique taxa and associated taxonomicinformation and two,
named lutaxa, a lookup from unique values in sppcol to matched taxonomy from
GBIF backbone.
}
\description{
Create taxonomy lookups
}
\examples{

}
