% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy_fixes}
\alias{taxonomy_fixes}
\title{Manual taxonomic fixes}
\format{
A data frame with 20 rows and 3
variables:
\describe{
\item{resolved_to}{Character. Taxa retrieved by \code{rgbif::name_backbone},
usually run within \code{envClean::make_taxa_taxonomy}}
\item{prefer}{Character. Preferred taxon to use instead of \code{resolved_to}.}
\item{note}{Character. Currently has mostly common names.}
...
}
}
\usage{
taxonomy_fixes
}
\description{
Use on results from \code{envClean::make_taxa_taxonomy}.
}
\keyword{datasets}
