% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy_overrides}
\alias{taxonomy_overrides}
\title{Manual taxonomic overrides}
\format{
A data frame with 13 rows and 4
variables:
\describe{
\item{original}{Character. Original names of taxa}
\item{prefer}{Character. Preferred taxon to use instead of any result resturned by the GBIF Backbone Taxonomy`.}
\item{note}{Character. Currently has mostly common names.}
...
}
}
\usage{
taxonomy_overrides
}
\description{
Can be used as argument to \code{envClean::make_taxa_taxonomy()}.
}
\keyword{datasets}
