% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_name_via_gbif.R
\name{try_name_via_gbif}
\alias{try_name_via_gbif}
\title{Attempt to find an unmatched scientific name using GBIF Backbone Taxonomy}
\usage{
try_name_via_gbif(name)
}
\arguments{
\item{name}{Character. Taxa name to search.}
}
\value{
If no match, NULL. If matched, a tibble ready for input to the
overrides argument of \code{make_taxonomy()}.
}
\description{
Useful in cases where the supplied scientific name does not find any match
in the taxonomic source being queried. This function queries the GBIF
Backbone Taxonomy for a supplied scientific name. The scientific
name from any successful match is then queried against
\code{galah::search_taxa()}.
}
\examples{
galah::search_taxa("Peziza vesiculosa Bull.: Fr.") # Homonym issue
try_name_via_gbif("Peziza vesiculosa Bull.: Fr.")
}
