% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cover.R
\name{add_cover}
\alias{add_cover}
\title{Generate best guess of cover for each taxa*context}
\usage{
add_cover(df, context = "cell", covCol = "useCover", smallcov = 0.009)
}
\arguments{
\item{df}{Dataframe with context, taxa and cover columns.}

\item{context}{Character. Name of columns defining context.}

\item{smallcov}{A small cover value assigned to any record for which there
is no available site, pca or taxa cover value.}
}
\value{
Dataframe with covCol removed and replaced with best guess cover in
column useCover
}
\description{
Generate best guess of cover for each taxa*context
}
