% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cover.R
\name{add_cover}
\alias{add_cover}
\title{Generate best guess of cover for each taxa * context}
\usage{
add_cover(
  df,
  context = "cell",
  env_prcomp,
  lucover,
  cover_cut_col = "cover_max",
  lucover_col = "cover_mid",
  small_cov = 0.009,
  remove_all_small = TRUE,
  fix_1to100 = TRUE,
  cov_func = "max"
)
}
\arguments{
\item{df}{Dataframe with context, taxa and cover columns.}

\item{context}{Character. Name of columns defining context.}

\item{env_prcomp}{List. Result from call to \code{envClean::env_pca()}}

\item{lucover}{Dataframe. Lookup from \code{cover_code} to numeric cover values}

\item{cover_cut_col}{Character. Name of column in lucover containing values
between 0 and 1 to use for assigning numeric cover values to a \code{cover_code}.
This allows the values appearing in the resulting tibble column \code{cover_adj}
to be continuous in 0 to 1.}

\item{lucover_col}{Character. Name of column in lucover containing values
to use for assigning numeric cover values to a \code{cover_code}. These do not
have to be between 0 and 1. These values are reflected in the \code{use_cover}
value in the resulting tibble.}

\item{small_cov}{A small cover value assigned to any record for which there
is no available site, pca or taxa cover value.}

\item{remove_all_small}{Logical. If \code{TRUE} (default) context(s) where all
taxa are assigned \code{small_cov} are removed.}

\item{fix_1to100}{Logical. Any values found in \code{cover} field of \code{df} that are
above 1 and less than or equal to 100 are divided by 100.}

\item{cov_func}{Function to summarise cover values for any context with more
than one cover value.}
}
\value{
Dataframe with taxa, context and the following 'cover' columns:
\code{cover_adj} a best guess proportion cover; \code{use_cover_code} the
\code{cover_code} values from \code{lucover} that correspond to the best guess value;
\code{use_cover} the \code{lucover_col} column that corresponds to the
best guess value.
}
\description{
Resulting 'best guess' cover values are generated from (in order): recorded
in the field; taxa within environmental pca groups (defined within
\code{env_prcomp}); and taxa across all its records. If more than one value is
available, \code{cov_func} is used to summarise to a single value.
}
