% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cover.R
\name{add_cover}
\alias{add_cover}
\title{Generate best guess of cover for each taxa*context}
\usage{
add_cover(df, context = "cell", covcol = "use_cover", smallcov = 0.009)
}
\arguments{
\item{df}{Dataframe with context, taxa and cover columns.}

\item{context}{Character. Name of columns defining context.}

\item{covcol}{Character. Name of column containing cover values.}

\item{smallcov}{A small cover value assigned to any record for which there
is no available site, pca or taxa cover value.}
}
\value{
Dataframe with covcol removed and replaced with best guess cover in
column use_cover
}
\description{
Generate best guess of cover for each taxa*context
}
