% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_taxa.R
\name{bin_taxa}
\alias{bin_taxa}
\title{Add \code{taxa} column}
\usage{
bin_taxa(df, taxa_col = "original_name", taxonomy)
}
\arguments{
\item{df}{Dataframe to clean, filter and tidy with respect to taxonomy.}

\item{taxa_col}{Character. Name of column with taxa.}

\item{taxonomy}{list with (at least) named elements \code{lutaxa} and \code{taxonomy}.
Usually resulting from call to \code{envClean::make_taxonomy()}.}
}
\value{
Dataframe with added column \code{taxa}
}
\description{
\code{taxa} column is the best match for the name in \code{taxa_col} based on the
GBIF Backbone Taxonomy.
}
