% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_text.R
\name{cleaning_text}
\alias{cleaning_text}
\title{Write a sentence describing change in taxa, records, visits and sites between
two cleaning steps}
\usage{
cleaning_text(prefix = "bio_", save_ends = TRUE, save_dir = "out", ...)
}
\arguments{
\item{prefix}{Character. Prefix used to name objects during the cleaning
process}

\item{save_ends}{Logical. Save the first and last data frames created by the
cleaning process?}

\item{save_dir}{Character. Path to save objects to if \code{save_ends == TRUE}}

\item{...}{Passed to \code{envClean::rec_vis_sit_tax()}}
}
\description{
Write a sentence describing change in taxa, records, visits and sites between
two cleaning steps
}
