% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{create_ind_status}
\alias{create_ind_status}
\title{Create indigenous status lookup}
\usage{
create_ind_status(df, sppcol = "taxa", indcol = "ind")
}
\arguments{
\item{df}{Dataframe with species column and column indicating indigenous
status.}

\item{sppcol}{Character. Name of column with species.}

\item{indcol}{Character. Name of column with indigenous status}
}
\value{
Dataframe with unique species and their indigenous status.
}
\description{
Create indigenous status lookup
}
