% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{create_lifeform}
\alias{create_lifeform}
\title{Get unique lifeform across species, perhaps including further context}
\usage{
create_lifeform(df, sppcol = "taxa", lfcol = "lifeform", context = NULL)
}
\arguments{
\item{df}{Dataframe with species column.}

\item{sppcol}{Character. Name of species column.}

\item{lfcol}{Character. Name of lifeform (id) column.}

\item{context}{Charcter or NULL. Set of columns that define a context within
which to generate lifeform.}
}
\value{
Dataframe with columns sppcol, visit col(s), lifeform
}
\description{
There are two tasks here:
}
\details{
\itemize{
\item find the most frequent lifeform assigned to a taxa (perhaps including
other context)
\item ensure there are no duplicates across the relevant context
}
}
