% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{effort_model}
\alias{effort_model}
\title{Model the effect of principal components axes on taxa richness.}
\usage{
effort_model(
  df,
  envprcomp,
  context = "cell",
  doIter = 1000,
  doChains = 3,
  threshold = 0.05
)
}
\arguments{
\item{df}{Dataframe. Cleaned data specifying context.}

\item{envprcomp}{Output from env_pca.}

\item{context}{Character. Column names that define context, usually a 'visit'
to a 'cell'.}

\item{doIter}{Numeric specifying the number of iterations to run for each
chain in rstan analysis.}

\item{doChains}{Numeric specifying the number of chains to run in rstan
analysis}

\item{threshold}{Numeric between 0 and 1 specifying the two-tailed threshold
above/below which richness is excessively above or below 'normal' for}
}
\value{

}
\description{
Model the effect of principal components axes on taxa richness.
}
