% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_prop}
\alias{filter_prop}
\title{Filter taxa recorded at less than x percent of visits}
\usage{
filter_prop(
  df,
  context = "cell",
  minsites = 15,
  keeptaxa = NULL,
  defaultper = 1
)
}
\arguments{
\item{df}{Dataframe with taxa and context}

\item{context}{Character. Column names that define context, usually a 'visit'
to a 'cell'.}

\item{minsites}{Absolute minimum sites at which a taxa should be recorded.}

\item{keeptaxa}{Character. taxa that should not be dropped. Used to set x
percent of sites.}

\item{defaultper}{If keeptaxa is NULL, what is the minimum percent of sites
at which a taxa should be recorded.}
}
\value{
df filtered to exclude taxa recorded at less than x percent of
visits.
}
\description{
Filter taxa recorded at less than x percent of visits
}
