% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_prop}
\alias{filter_prop}
\title{Filter taxa recorded at less than x percent of visits}
\usage{
filter_prop(df, context = "cell", min_sites = 15, keep = NULL, default_per = 1)
}
\arguments{
\item{df}{Dataframe with taxa and context}

\item{context}{Character. Column names that define context, usually a 'visit'
to a 'cell'.}

\item{min_sites}{Absolute minimum sites at which a taxa should be recorded.}

\item{keep}{Character. taxa that should not be dropped. Used to set x
percent of sites.}

\item{default_per}{If keeptaxa is NULL, what is the minimum percent of sites
at which a taxa should be recorded.}
}
\value{
df filtered to exclude taxa recorded at less than x percent of
visits.
}
\description{
Filter taxa recorded at less than x percent of visits
}
