% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_spat_rel}
\alias{filter_spat_rel}
\title{Filter data frame to specified spatial reliability}
\usage{
filter_spat_rel(df, distcol = "maxDist", dist, context, dfrel)
}
\arguments{
\item{df}{Dataframe.}

\item{distcol}{Character. Name of the column containing the spatial reliability.}

\item{dist}{Numeric. In units of the maxDist column.}

\item{context}{Character. column names defining the context.}

\item{dfrel}{Dataframe. Lookup from reliability id to reliability distance.}
}
\value{
Dataframe with records of greater than dist filtered. Filtering is done
at 'visit' level.
}
\description{
Filter data frame to specified spatial reliability
}
