% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_spat_rel}
\alias{filter_spat_rel}
\title{Filter data frame to specified spatial reliability}
\usage{
filter_spat_rel(
  df,
  dist_col = "max_dist",
  default = Inf,
  dist = 50,
  context,
  df_rel
)
}
\arguments{
\item{df}{Dataframe.}

\item{dist_col}{Character. Name of the column containing the spatial
reliability.}

\item{default}{Numeric. Default value for \code{dist_col} to use if \code{dist_col} is
\code{NA}. Useful when importing reliable data with no associated reli}

\item{dist}{Numeric. Filter spatial reliability greater than this value. In
the same units as \code{dist_col}.}

\item{context}{Character. column names defining the context.}

\item{df_rel}{Dataframe. Lookup from reliability id to reliability distance.}
}
\value{
Dataframe with records of greater than dist filtered. Filtering is
done at \code{context} level.
}
\description{
Filter data frame to specified spatial reliability
}
