% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Clean/Tidy to one row per Taxa*Visit}
\usage{
filter_taxa(
  df,
  sppcol = "SPECIES",
  visit = visitcols,
  extracols = NULL,
  targetRank = "Species",
  doCover = FALSE,
  doLifeform = FALSE
)
}
\arguments{
\item{df}{Dataframe to reduce.}

\item{sppcol}{Character. Name of column with species.}

\item{visit}{Character. Name of columns defining a visit.}

\item{extracols}{Character. Name of any extra columns to keep.}
}
\value{
Dataframe with columns Taxa, visit column(s) and, if used, extracols.
}
\description{
Includes running of taxa_taxonomy()
}
