% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Clean/Tidy to one row per taxa*Visit}
\usage{
filter_taxa(
  df,
  taxa_col = "original_name",
  context,
  extra_cols = NULL,
  target_rank = "species",
  do_cov = FALSE,
  do_life = FALSE,
  do_ind = FALSE,
  lucov = NULL,
  lulife = NULL,
  ...
)
}
\arguments{
\item{df}{Dataframe to reduce.}

\item{taxa_col}{Character. Name of column with taxa.}

\item{context}{Character. Name of columns defining context.}

\item{extra_cols}{Character. Name of any extra columns to keep.}

\item{target_rank}{Character. Name of level in taxonomic hierarchy that names
should be retrieved from, if possible.}

\item{do_cov}{Logical. Should cover (needs to be supplied in df) be appended
to output.}

\item{do_life}{Logical. Should lifeform (needs to be supplied in df) be
appended to output.}

\item{do_ind}{Logical. Should ind (needs to be supplied in df) be
appended to output.}

\item{lucov}{Dataframe lookup for cover.}

\item{lulife}{Dataframe lookup for lifeform.}

\item{...}{Passed to \verb{\link{envClean}{make_taxa_taxonomy()}}.}
}
\value{
Dataframe with columns taxa, visit column(s) and, if used, extracols.
}
\description{
Includes running of taxa_taxonomy(), if lutaxa does not already exist.
}
