% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_taxa.R, R/filter_taxa.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Add \code{taxa} column}
\usage{
filter_taxa(
  df,
  taxa_col = "original_name",
  context,
  extra_cols = NULL,
  required_rank = "species",
  do_cov = FALSE,
  do_life = FALSE,
  lucov = NULL,
  lulife = NULL,
  taxonomy
)

filter_taxa(
  df,
  taxa_col = "original_name",
  context,
  extra_cols = NULL,
  required_rank = "species",
  do_cov = FALSE,
  do_life = FALSE,
  lucov = NULL,
  lulife = NULL,
  taxonomy
)
}
\arguments{
\item{df}{Dataframe to clean, filter and tidy with respect to taxonomy.}

\item{taxa_col}{Character. Name of column with taxa.}

\item{context}{Character. Name of columns defining context.}

\item{extra_cols}{Character. Name of any extra columns to keep.}

\item{required_rank}{Character. Default is 'species'. What level of the
taxonomic hierarchy is required for results. Records above this level are
filtered.}

\item{do_cov}{Logical. Should cover (needs to be supplied in df) be appended
to output.}

\item{do_life}{Logical. Should lifeform (needs to be supplied in df) be
appended to output.}

\item{lucov}{Dataframe lookup for cover.}

\item{lulife}{Dataframe lookup for lifeform.}

\item{taxonomy}{list with (at least) named elements \code{lutaxa} and \code{taxonomy}.
Usually resulting from call to \code{envClean::make_taxonomy()}.}

\item{do_ind}{Logical. Should ind (needs to be supplied in df) be
appended to output.}
}
\value{
Dataframe with added column \code{taxa}

Dataframe with columns taxa, context and, possibly, extracols
, lifeform and cover
}
\description{
\code{taxa} column is the best match for the name in \code{taxa_col} based on the
GBIF Backbone Taxonomy.

Deprecated in favour of the simpler workflow involving make_taxonomy then
bin_taxa (along with other bins, say, geographic and temporal). Required
attributes are then added to each taxa; context and/or taxa + context before
reducing to distinct context and associated attributes.
}
