% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Clean/Tidy to one row per taxa*Visit}
\usage{
filter_taxa(
  df,
  taxa_col = "original_name",
  context,
  extra_cols = NULL,
  do_cov = FALSE,
  do_life = FALSE,
  do_ind = FALSE,
  lucov = NULL,
  lulife = NULL,
  taxonomy = NULL,
  target_rank = "species",
  ...
)
}
\arguments{
\item{df}{Dataframe to reduce.}

\item{taxa_col}{Character. Name of column with taxa.}

\item{context}{Character. Name of columns defining context.}

\item{extra_cols}{Character. Name of any extra columns to keep.}

\item{do_cov}{Logical. Should cover (needs to be supplied in df) be appended
to output.}

\item{do_life}{Logical. Should lifeform (needs to be supplied in df) be
appended to output.}

\item{do_ind}{Logical. Should ind (needs to be supplied in df) be
appended to output.}

\item{lucov}{Dataframe lookup for cover.}

\item{lulife}{Dataframe lookup for lifeform.}

\item{taxonomy}{list with named elements \code{lutaxa} and \code{taxa_taxonomy}.
Usually resulting from call to \code{make_taxa_taxonomy}.}

\item{target_rank}{Character. Default is 'species'. At what level of the
taxonomic hierarchy are results desired.}

\item{...}{Passed to \code{envClean::make_taxa_taxonomy}.}
}
\value{
Dataframe with columns taxa, context and, if used, extracols.
}
\description{
Includes running of taxa_taxonomy(), if lutaxa does not already exist.
}
