% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_text_col}
\alias{filter_text_col}
\title{Filter a column}
\usage{
filter_text_col(df, filt_col, filt_text, df_join = NULL)
}
\arguments{
\item{df}{Dataframe with column to filter.}

\item{filt_col}{Character. Name of column to filter.}

\item{filt_text}{Character. Text(s) to filter from df.}

\item{df_join}{Optional dataframe. Joined to df before filter. No names from
dfJoin are returned.}
}
\value{
Filtered dataframe with same names as df
}
\description{
Filter a column
}
