% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{filter_text_col}
\alias{filter_text_col}
\title{Filter a column}
\usage{
filter_text_col(df, filtcol, filttext, dfjoin = NULL)
}
\arguments{
\item{df}{Dataframe with column to filter.}

\item{filtcol}{Character. Name of column to filter.}

\item{filttext}{Character. Text(s) to filter from df.}

\item{dfjoin}{Optional dataframe. Joined to df before filter. No names from
dfJoin are returned.}
}
\value{
Filtered dataframe with same names as df
}
\description{
Filter a column
}
