% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{get_gbif_tax}
\alias{get_gbif_tax}
\title{Get taxonomy from \href{https://www.gbif.org/dataset/d7dddbf4-2cf0-4f39-9b2a-bb099caae36c}{GBIF Backbone Taxonomy}.}
\usage{
get_gbif_tax(
  df,
  taxa_col = "original_name",
  out_file = tempfile(),
  king_type = "Plantae",
  do_common = FALSE,
  target_rank = "species"
)
}
\arguments{
\item{df}{Dataframe with column of species names to resolve}

\item{taxa_col}{Character. Name of column with species names}

\item{out_file}{Character. Path to save results to. Saving is iterative as
retrieving names can take some time.}

\item{king_type}{Character. Kingdom type (i.e. Plantae, Animalia etc.)}

\item{do_common}{Logical. If TRUE, also get common names. Takes much longer.}

\item{target_rank}{Character. Default is 'Species'. At what level of the
taxonomic hierarchy are results desired.}
}
\value{
Dataframe of unique species names, other columns defining
taxonomic hierarchy and, optionally, common names.
}
\description{
Retrieve accepted name taxonomy and retrieve taxonomic hierarchy for a df
with a column of taxonomic names
}
\examples{
get_gbif_tax(data.frame(spp = "Eucalyptus viminalis"), taxa_col = "spp")


}
