% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomy.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Get GBIF backbone taxonomy}
\usage{
get_taxonomy(
  df,
  taxa_col = "original_name",
  taxonomy_file = tempfile(),
  do_common = FALSE,
  remove_taxa = c("BOLD:", "dead", "unverified", "annual herb", "annual grass", "\\\\?"),
  remove_strings = c("\\\\sx\\\\s.*|\\\\sX\\\\s.*"),
  ...
)
}
\arguments{
\item{df}{Dataframe with taxa column.}

\item{taxa_col}{Character. Name of column with taxa names}

\item{taxonomy_file}{Character. Path to save results to.}

\item{remove_taxa}{Character. Regular expressions to be matched. These will
be filtered before searching.}

\item{remove_strings}{Character. Regular expressions to be matched. These
will be removed from the string before searching.}

\item{...}{Arguments passed to \code{rgbif::name_backbone_checklist()}.}
}
\value{
Dataframe. Results from \code{envClean::get_gbif_tax()}. Tweaked by column
\code{rank} being lowercase and factor as per \code{envClean::lurank}.
}
\description{
Only queries GBIF for taxa not already in \code{taxonomy_file}.
}
\details{
Common (vernacularName) no longer supported here. Use \code{get_gbif_common()} on
a downstream result. It may be helpful to keep a usageKey through the
cleaning process for use in getting common names. Part of the reason for
removing that functionality here was the ambiguity of which key to use,
particularly around species vs subspecies.
}
