% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lufilter}
\alias{lufilter}
\title{Dataframe of filtering steps}
\format{
A data frame with 16 rows and 3
variables:
\describe{
\item{filter}{Character short name for filtering step.}
\item{desc}{Phrase describing filtering step.}
\item{order}{Suggested order}
...
}
}
\usage{
lufilter
}
\description{
Dataframe of filtering steps
}
\keyword{datasets}
