% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lurank}
\alias{lurank}
\title{Dataframe of taxonomic ranks}
\format{
A data frame with 10 rows and 2
variables:
\describe{
\item{rank}{Taxonomic rank.}
\item{sort}{Numeric. Can be used directly to sort taxonomic ranks but is
also used to create the ordered factor rank.}
...
}
}
\usage{
lurank
}
\description{
A dataset listing taxonomic rank as an ordered factor.
}
\keyword{datasets}
