% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_effort_mod.R
\name{make_effort_mod}
\alias{make_effort_mod}
\title{Distribution of credible values for taxa richness.}
\usage{
make_effort_mod(
  df,
  context = "cell",
  cat_cols = NULL,
  threshold_lo = 0.05/2,
  threshold_hi = 0.05/2,
  use_family = rstanarm::neg_binomial_2(),
  ...
)
}
\arguments{
\item{df}{Dataframe. Cleaned data specifying context.}

\item{context}{Character. Column names that define context, usually a 'visit'
to a 'cell'.}

\item{cat_cols}{Character. Name of column(s) (0 to 2) specifying the
categorical variables to model. Usually a taxonomic level (say, class) and a
geographic level (say, IBRA Region). If NULL (default), model is \code{y ~ 1}.}

\item{threshold_lo, threshold_hi}{Numeric between 0 and 1 specifying the
threshold above/below which richness is excessively above or below 'normal'
and should be filtered.}

\item{...}{Other arguments passed to \code{rstanarm::stan_glm} (e.g. chains, iter).}

\item{family}{Passed to \code{rstanarm::stan_glm}.}
}
\value{
List of model outputs.
}
\description{
Model the effect of between 0 and 2 categorical variables on taxa richness
within a context. Highlight, based on thresholds, contexts where taxa
richness is 'excessively' high or low.
}
