% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{make_taxa_taxonomy}
\alias{make_taxa_taxonomy}
\title{Make taxonomy lookups}
\usage{
make_taxa_taxonomy(
  df,
  taxa_col = "original_name",
  lifespan_col = NULL,
  ind_col = NULL,
  poor_filt = c("dead", "unverified"),
  save_luGBIF = tempfile(),
  king = "Plantae"
)
}
\arguments{
\item{df}{Dataframe with species column.}

\item{taxa_col}{Name of column with species.}

\item{lifespan_col}{Character. Optional name of column containing lifespan
information.}

\item{ind_col}{Character. Optional name of columns containing indigenous
status of taxa in taxa_col.}

\item{poor_filt}{Character. Any taxa names to grep out of the species column.
(e.g. c("annual form", "unverified")).}

\item{save_luGBIF}{Character. Path to file containing desired taxonomy to use.
This is usually the output from gfbif_tax(). If this does not exist it will
be created (as tempfile) by get_gbif_tax().}

\item{king}{Character. Kingdom to search preferentially in GBIF Taxonomy
Backbone}
}
\value{
Dataframe with applied taxonomy from GBIF Taxonony Backbone. Also,
two dataframes are returned to the global environment. One, named
taxaTaxonomy, with unique taxa and associated taxonomicinformation and two,
named lutaxa, a lookup from unique values in taxa_col to matched taxonomy from
GBIF backbone.
}
\description{
Make taxonomy lookups
}
\examples{

}
