% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxonomy.R
\name{make_taxonomy}
\alias{make_taxonomy}
\title{Make taxonomy lookups}
\usage{
make_taxonomy(
  df,
  taxa_col = "original_name",
  taxonomy_file,
  target_rank = "species",
  limit = TRUE,
  ...
)
}
\arguments{
\item{df}{Dataframe with taxa column.}

\item{taxa_col}{Character. Name of column with taxa names}

\item{target_rank}{Character. Default is 'species'. At what level of the
taxonomic hierarchy are results desired.}

\item{limit}{Logical. If true (default), the output taxonomy will be limited
to the input names in \code{taxa_col}. Otherwise, any taxa in \code{out_file} will be
returned.}

\item{...}{Passed to \code{get_taxonomy()}}

\item{out_file}{Character. Path to save results to.}

\item{do_common}{Logical. If true, an attempt will be made to find a common
name for each taxa at \code{target_rank}.}
}
\value{
named list with elements:
\item{lutaxa}{Dataframe. For each \code{original_name} a taxa to use.}
\item{taxonomy}{Dataframe. For each taxa a row of taxonomic hierarchy and
matching gbif usageKeys to \code{target_rank} level}
\item{common}{If \code{do_common} a dataframe with common name for each
taxa in \code{taxonomy}.}
}
\description{
Make taxonomy lookups
}
