% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxonomy.R, R/test.R
\name{make_taxonomy}
\alias{make_taxonomy}
\title{Make taxonomy lookups}
\usage{
make_taxonomy(
  df,
  taxa_col = "original_name",
  taxonomy_file,
  target_rank = "species",
  limit = TRUE,
  ...
)

make_taxonomy(
  df,
  taxa_col = "original_name",
  taxonomy_file,
  target_rank = "species",
  limit = TRUE,
  ...
)
}
\arguments{
\item{df}{Dataframe with taxa column.}

\item{taxa_col}{Character. Name of column with taxa names}

\item{taxonomy_file}{Character. Path to results from \code{envFunc::get_taxonomy()}}

\item{target_rank}{Character. Default is 'species'. At what level of the
taxonomic hierarchy are results desired.}

\item{limit}{Logical. If true (default), the output taxonomy will be limited
to the input names in \code{taxa_col}. Otherwise, any taxa in \code{out_file} will be
returned.}

\item{...}{Passed to \code{get_taxonomy()}}

\item{out_file}{Character. Path to save results to.}
}
\value{
named list with elements:
\item{lutaxa}{Dataframe. For each unique name in \code{taxa_col}, the best
\code{taxa} to use (taking into account \code{target_rank})}
\item{taxonomy}{Dataframe. For each \code{taxa}in \code{lutaxa} a row of taxonomic
hierarchy and matching gbif usageKeys}

named list with elements:
\item{lutaxa}{Dataframe. For each \code{original_name} a taxa to use.}
\item{taxonomy}{Dataframe. For each taxa a row of taxonomic hierarchy and
matching gbif usageKeys to \code{target_rank} level}
}
\description{
Make taxonomy lookups

Make taxonomy lookups
}
