% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxonomy.R
\name{make_taxonomy}
\alias{make_taxonomy}
\title{Make taxonomy lookups}
\usage{
make_taxonomy(
  df,
  taxa_col = "original_name",
  out_file = tempfile(),
  target_rank = "species",
  do_common = TRUE,
  limit = TRUE,
  lifespan_col = NULL,
  ind_col = NULL,
  remove_taxa = c("BOLD:", "dead", "unverified", "annual herb", "annual grass", "\\\\?"),
  remove_strings = NULL,
  ...
)
}
\arguments{
\item{df}{Dataframe with taxa column.}

\item{taxa_col}{Character. Name of column with taxa names}

\item{out_file}{Character. Path to save results to.}

\item{target_rank}{Character. Default is 'species'. At what level of the
taxonomic hierarchy are results desired.}

\item{do_common}{Logical. If true, an attempt will be made to find a common
name. Common name matches are attempted at \code{target_rank} (or higher taxonomic
rank if no taxa was matched at \code{target_rank} for the name supplied in
\code{taxa_col}).
\code{envClean::get_gbif_tax}.}

\item{limit}{Logical. If true, the output taxonomy will be limited to the
input names in \code{taxa_col}. Otherwise, any taxa in \code{out_file} will be
returned.}

\item{lifespan_col}{Character. Optional name of column containing lifespan
information.}

\item{ind_col}{Character. Optional name of columns containing indigenous
status of taxa in taxa_col.}

\item{remove_taxa}{Character. Regular expressions to be matched. These will
be filtered before searching.}

\item{remove_strings}{Character. Regular expressions to be matched. These
will be removed from the string before searching.}

\item{...}{Arguments passed to \code{rgbif::name_backbone_checklist()}.}
}
\value{
named list with elements:
\item{raw}{Dataframe. Results from \code{envClean::get_gbif_tax()}}
\item{lutaxa}{dataframe. A lookup from unique values in \code{taxa_col} to
matched taxonomy from GBIF backbone}
\item{taxonomy}{dataframe with unique taxa and associated taxonomic
information}
\item{common}{If \code{do_common} a dataframe with unique taxa and associated
common name}
}
\description{
Make taxonomy lookups
}
