% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envClean.R
\name{taxa_taxonomy}
\alias{taxa_taxonomy}
\title{Create taxonomy lookups}
\usage{
taxa_taxonomy(
  df,
  sppCol = "SPECIES",
  lifespanCol = NULL,
  poorTaxFilt = speciesFilter,
  luTaxFile = "out/luGBIF.feather",
  king = "Plantae"
)
}
\arguments{
\item{df}{Dataframe with species column.}

\item{sppCol}{Name of column with species.}

\item{poorTaxFilt}{Character. Any text to grep out of the species column.
(e.g. c("dead", "unknown")).}

\item{luTaxFile}{Character. Path to file containing desired taxonomy to use.
This is usually the output from gfbif_tax(). If this does not exist it will
be created by gbif_tax().}

\item{king}{Character. Kingdom to search preferentially in GBIF Taxonomy
Backbone}
}
\value{
Dataframe with applied taxonomy from GBIF Taxonony Backbone. Also,
two dataframes are returned to the global environment. One, named
taxaTaxonomy, with unique taxa and associated taxonomicinformation and two,
named luTaxa, a lookup from unique values in sppCol to matched taxonomy from
GBIF backbone.
}
\description{
Create taxonomy lookups
}
\examples{

}
