% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy_overrides}
\alias{taxonomy_overrides}
\title{Manual taxonomic overrides}
\format{
A data frame with 25 rows and 5
variables:
\describe{
\item{original_name}{Character. Original names of taxa}
\item{taxa_to_search}{Character. Try searching for this taxa (often the
same as original_name)}
\item{use_}{Character. Any number of \code{use_} columns, such as \code{use_species}
and/or \code{use_subspecies}. If no match is found for \code{taxa_to_search} these
names will be used at the rank they imply}
\item{note}{Usually a rationale for the inclusion}
...
}
}
\usage{
taxonomy_overrides
}
\description{
Can be used as argument to \code{envClean::make_taxa_taxonomy()}.
}
\keyword{datasets}
